Issue-43: When --format is specified with the log method, return everything.
This commit addresses Issue 43, which identifies the limitation in G::W that
there is no way to parse out each log entry for custom formats.

The solution is to provide a G::W::Log::Custom wrapper that returns all of
the output from the log method in the specified format with no attempt
to parse this output into individual records.

If the caller of the log method specifies the format, then they are responsible
for parsing the output.

There is the tempation here to all a parse subroutine reference to be passed,
but I would strongly encourage that we not get into the business of parsing
customized git log formats.